requirement.txt
# Backend
fastapi==0.111.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1

# IA / Embeddings / RAG
cohere==4.0.0
chromadb==0.4.8

# PDF Processing
pdfplumber==0.9.0

# Frontend
streamlit==1.29.0
plotly==5.16.1

# Utilities
pandas==2.1.1
tqdm==4.66.1
